Bookstore assignment
For this assignment 2, I created a bookshop cart that adds the number of items into the cart, calculate the total, the quantity of each item and then gives you a discount if you buy mor4e bulk items, gives you discount to the customers with store membership when the total is done. The bookstore cart also gives a discount to sudden items when you buy two or more of it. I created JUnit test for Item, ItemOrder, and Cart to test them. Using BigDecimal to round to the precise Integer value in the hashmap. The program was helpful in showing during hashmap was better than using ArrayList when storing the items into the cart.